+++ docs/User/ProjectRelated/Changelog.md
@@ -4,10 +4,9 @@
 ## 0.9.0 (Unreleased)
 
 * Remove support for GHC 8.4.*
-* The unboxed arrays now require `Unbox` instance along with `Storable` for the
-  stored type. The `Unbox` typeclass can be found in `Streamly.Data.Unbox`.
 * New modules for unboxed mutable & unboxed immutable arrays are added.
-* A new module dedicated for unboxing is added.
+* The unboxed arrays now require `Unbox` instead of `Storable` for the stored
+  type. The `Unbox` typeclass can be found in `Streamly.Data.Array.Unboxed`.
 
 * Following APIs have been renamed in Streamly.Network.Socket module:
     readWithBufferOf -> readWith